# Summarizing performance metrics and resource usage results