Takes in the saved results from the PyTorch profiler and exports reports (stdout and csv) to compare across experiments.

Example usage:

```
from ludwig.benchmarking.summarize import summarize_metrics

metric_diffs, resource_usage_diffs = summarize_metrics(
    "/Users/projects/ludwig/benchmarking/configs/benchmarking_config.yaml",
    "local_test_1",
    "local_test_2",
    "s3://some_bucket/benchmarks/")

for experiment_metric_diff in metric_diffs:
    print(experiment_metric_diff.to_string())

for experiment_resource_diff in resource_usage_diffs:
    for tag_diff in experiment_resource_diff:
        print(tag_diff.to_string())
```

This will print an export to CSV the following table for any function flagged to be profiled. In the current example, we profile the function `ludwig.api.LudwigModel.evaluate`.
```
Metrics for dataset: ames_housing
Output Feature Name  Metric Name             local_test_1  local_test_2  Diff          Diff Percentage
SalePrice            mean_absolute_error     180309.547    180308.281    -1.266        -0.001
SalePrice            mean_squared_error      38554296320.0 38553845760.0 -450560.0     -0.001
SalePrice            r2                      -5.38         -5.38         0.0           -0.001
SalePrice            root_mean_squared_error 196352.484    196351.328    -1.156        -0.001
SalePrice            root_mean_squared_percentage_error 1.0           1.0           -0.0          -0.001

Resource usage for: LudwigModel.evaluate
Metric Name                    local_test_1         local_test_2         Diff                 Diff Percentage
average_cpu_memory_usage       338.08 Mb            660.02 Mb            321.94 Mb            95.226
average_cpu_utilization        90.9                 810.9732558139535    720.0732558139536    792.16
cpu_memory_available           6.36 Gb              990.04 Gb            983.68 Gb            15459.893
disk_footprint                 32768                17068032             17035264             51987.5
max_cpu_memory_usage           338.08 Mb            669.75 Mb            331.68 Mb            98.106
max_cpu_utilization            98.5                 1958.1               1859.6               1887.919
num_cpu                        8                    255                  247                  3087.5
num_runs                       1                    1                    0                    0.0  
torch_average_cpu_memory_usage 182.36 Kb            182.36 Kb            -0.03467944375006482 b -0.0 
torch_cpu_time                 231.868ms            128.488s             128.256s             55314.392
torch_cpu_time_total           231.868ms            128.488s             128.256s             55314.392
torch_cuda_time                0.000us              0.000us              0.000us              inf  
torch_cuda_time_total          0.000us              0.000us              0.000us              inf  
torch_max_cpu_memory_usage     467.41 Kb            467.41 Kb            0 b                  0.0  
torch_self_cpu_time_total      201.276ms            3.229s               3.028s               1504.485
torch_self_cuda_time_total     0.000us              0.000us              0.000us              inf  
total_execution_time           235.431ms            128.492s             128.256s             54477.191
```